Add move camera button to multi display activity and secondary screen activity.
