+++ app/src/main/java/com/mapbox/maps/testapp/examples/SecondaryDisplayActivity.kt
@@ -1,0 +1,67 @@
+package com.mapbox.maps.testapp.examples
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.mapbox.geojson.Point
+import com.mapbox.maps.CameraOptions
+import com.mapbox.maps.Style
+import com.mapbox.maps.plugin.animation.MapAnimationOptions
+import com.mapbox.maps.plugin.animation.flyTo
+import com.mapbox.maps.testapp.R
+import kotlinx.android.synthetic.main.activity_multi_display.*
+import kotlinx.android.synthetic.main.activity_simple_map.mapView
+
+/**
+ * Example of displaying a map on a secondary display.
+ */
+class SecondaryDisplayActivity : AppCompatActivity() {
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_secondary_display)
+    mapView.getMapboxMap().loadStyleUri(Style.MAPBOX_STREETS)
+    moveCameraButton.setOnClickListener {
+      moveCamera()
+    }
+  }
+
+  private fun moveCamera() {
+    val cameraOption = CameraOptions.Builder()
+      .center(HELSINKI)
+      .zoom(ZOOM)
+      .build()
+
+    mapView.getMapboxMap().flyTo(
+      cameraOption,
+      MapAnimationOptions.mapAnimationOptions {
+        duration = DURATION
+      }
+    )
+  }
+
+  override fun onStart() {
+    super.onStart()
+    mapView.onStart()
+  }
+
+  override fun onStop() {
+    super.onStop()
+    mapView.onStop()
+  }
+
+  override fun onLowMemory() {
+    super.onLowMemory()
+    mapView.onLowMemory()
+  }
+
+  override fun onDestroy() {
+    super.onDestroy()
+    mapView.onDestroy()
+  }
+
+  companion object {
+    private val HELSINKI = Point.fromLngLat(24.9384, 60.1699)
+    private const val ZOOM = 14.0
+    private const val DURATION = 3000L
+  }
+}
