--- app/src/main/java/com/mapbox/maps/testapp/examples/SecondaryDisplaySimpleMapActivity.kt
@@ -1,39 +1,0 @@
-package com.mapbox.maps.testapp.examples
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.mapbox.maps.Style
-import com.mapbox.maps.testapp.R
-import kotlinx.android.synthetic.main.activity_simple_map.*
-
-/**
- * Example of displaying a map.
- */
-class SecondaryDisplaySimpleMapActivity : AppCompatActivity() {
-
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    setContentView(R.layout.activity_simple_map)
-    mapView.getMapboxMap().loadStyleUri(Style.MAPBOX_STREETS)
-  }
-
-  override fun onStart() {
-    super.onStart()
-    mapView.onStart()
-  }
-
-  override fun onStop() {
-    super.onStop()
-    mapView.onStop()
-  }
-
-  override fun onLowMemory() {
-    super.onLowMemory()
-    mapView.onLowMemory()
-  }
-
-  override fun onDestroy() {
-    super.onDestroy()
-    mapView.onDestroy()
-  }
-}
