+++ app/src/main/res/layout/activity_debug.xml
@@ -74,7 +74,7 @@
             android:text="@string/debug_depth_buffer" />
 
         <Button
-            android:id="@+id/button"
+            android:id="@+id/displayOnSecondDisplayButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/debug_button" />
