+++ app/src/main/res/layout/activity_empty_fab.xml
@@ -8,7 +8,7 @@
         android:layout_height="match_parent" />
 
     <Button
-        android:id="@+id/button"
+        android:id="@+id/displayOnSecondDisplayButton"
         android:layout_width="match_parent"
         android:layout_height="58dp"
         android:text="REPLACE WITH EMPTY FRAGMENT"
