+++ app/src/main/res/values/strings.xml
@@ -120,5 +120,6 @@
     <string name="filter">Filter</string>
     <string name="animate_symbol">Animate symbol</string>
     <string name="display_the_second_screen">Display map in the second screen</string>
+    <string name="move_camera">Move Camera</string>
 
 </resources>
